// ehif-dbi/src/main/frontend/src/app/api/models/parish-document-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ParishionerEmbeddedDto } from '../models/parishioner-embedded-dto';
import { PriestEmbeddedDto } from '../models/priest-embedded-dto';
export interface ParishDocumentDto {
  foundedYear?: number;
  id?: string;
  location?: string;
  name?: string;
  parishioners?: Array<ParishionerEmbeddedDto>;
  priests?: Array<PriestEmbeddedDto>;
}